+++ src/components/header/Header.tsx
@@ -1,15 +1,12 @@
-import { NavLinks } from './components/NavLinks';
-import { NavLink } from 'react-router-dom';
 import styles from './Header.module.scss';
-import { messages } from './messages';
-import { PagePath } from '@/router/enums';
+import { Logo } from './components/logo/Logo';
+import { NavLinks } from './components/nav-links/NavLinks';
 
 export const Header = () => {
   return (
     <header className={styles.container}>
-      <NavLink to={PagePath.root} className={styles.title}>
-        {messages.appTitle}
-      </NavLink>
+      <Logo />
+
       <NavLinks />
     </header>
   );
