+++ src/components/header/components/logo/Logo.module.scss
@@ -1,0 +1,9 @@
+.title {
+  text-decoration: none;
+  color: var(--clr-line);
+
+  &:hover {
+    transform: scale(1.03);
+    color: var(--clr-hover);
+  }
+}
